feat, breaks (ViewRequestsActivity.java) : Imported the necessary APIs for the email feature. And added the method calls to the relevant onClick methods
It's being refactored right now, so it can't be helped. But currently, the request returns a userWrapped object, and the email method expects a User.
Could be fixed by creating a new User from the WrappedUser or by refactoring the sendEmail function, but since the userWrapped is going to be phased out soon anyways, I didn't see the point.